pyproject.toml
[project]
name = "servipal-backend"
version = "0.1.0"
description = "ServiPal: Multi-vendor lifestyle application backend"
dependencies = [
    "celery>=5.6.3",
    "exponent-server-sdk==2.2.0",
    "fastapi[standard]==0.128.0",
    "logfire[fastapi]==4.19.0",
    "postgrest==2.27.2",
    "pycryptodome>=3.23.0",
    "pydantic-settings==2.12.0",
    "pytest==9.0.2",
    "redis==7.1.0",
    "rq==2.6.1",
    "sentry-sdk[fastapi]==2.49.0",
    "structlog==25.5.0",
    "supabase==2.27.2",
    "uvicorn==0.40.0",
]

[dependency-groups]
dev = [
    "pytest-asyncio==1.3.0",
]
